requirements.txt
anyio==4.0.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
et-xmlfile==1.1.0
Flask==2.3.1
google-api-core==2.14.0
google-auth==2.23.4
google-cloud-bigquery==3.13.0
google-cloud-core==2.3.3
google-cloud-pubsub==2.18.4
google-cloud-storage==2.13.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
grpc-google-iam-v1==0.12.7
grpcio==1.59.3
grpcio-status==1.59.3
gunicorn==20.1.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.26.2
openpyxl==3.1.2
packaging==23.2
pandas==1.2.4
pluggy==1.3.0
proto-plus==1.22.3
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pytest==7.4.3
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
urllib3==2.1.0
Werkzeug==3.0.1
